experiment overlap schedule
