+++ python/sglang/srt/managers/schedule_batch.py
@@ -1511,6 +1511,25 @@ class ModelWorkerBatch:
     # If set, the output of the batch contains the hidden states of the run.
     capture_hidden_mode: CaptureHiddenMode = None
 
+    positions: Optional[torch.Tensor] = None
+    batch_size: Optional[int] = None
+    extend_start_loc: Optional[torch.Tensor] = None
+
+    # HPU-specific fields
+    page_size: Optional[int] = None
+    attn_bias: Optional[torch.Tensor] = None
+    seq_pos: Optional[torch.Tensor] = None
+    seq_idx: Optional[torch.Tensor] = None
+    valid_seq_len: Optional[torch.Tensor] = None
+    extend_seq_lens_padded: Optional[torch.Tensor] = None
+    real_batch_size: Optional[int] = None
+    use_contiguous_pa: Optional[bool] = None
+    block_list: Optional[torch.Tensor] = None
+    block_mapping: Optional[torch.Tensor] = None
+    block_groups: Optional[torch.Tensor] = None
+    block_usage: Optional[torch.Tensor] = None
+    block_scales: Optional[torch.Tensor] = None
+
 
 @triton.jit
 def write_req_to_token_pool_triton(
