+++ python/sglang/srt/managers/scheduler_output_processor_mixin.py
@@ -5,13 +5,15 @@ from typing import TYPE_CHECKING, List, Optional, Tuple, Union
 from sglang.srt.layers.logits_processor import LogitsProcessorOutput
 from sglang.srt.managers.io_struct import BatchEmbeddingOut, BatchTokenIDOut
 from sglang.srt.managers.schedule_batch import BaseFinishReason, Req, ScheduleBatch
-
+from sglang.srt.utils import is_hpu
 if TYPE_CHECKING:
     from sglang.srt.managers.scheduler import (
         EmbeddingBatchResult,
         GenerationBatchResult,
         ScheduleBatch,
     )
+
+_is_hpu = is_hpu()
 
 
 class SchedulerOutputProcessorMixin:
@@ -206,11 +208,10 @@ class SchedulerOutputProcessorMixin:
                 continue
 
             if self.enable_overlap and req.finished():
-                raise NotImplementedError("Not implemented for overlap")
                 # Free the one extra delayed token
                 if self.page_size == 1:
                     self.token_to_kv_pool_allocator.free(batch.out_cache_loc[i : i + 1])
-                else:
+                elif not _is_hpu:
                     # Only free when the extra token is in a new page
                     if (
                         len(req.origin_input_ids) + len(req.output_ids) - 1
@@ -218,6 +219,9 @@ class SchedulerOutputProcessorMixin:
                         self.token_to_kv_pool_allocator.free(
                             batch.out_cache_loc[i : i + 1]
                         )
+                else:
+                    # TODO: review this to make sure it is correct
+                    pass
                 continue
 
             if batch.spec_algorithm.is_none():
