+++ python/sglang/srt/model_executor/forward_batch_info.py
@@ -38,6 +38,7 @@ import torch
 import triton
 import triton.language as tl
 import os
+import math
 
 from sglang.srt.layers.rotary_embedding import MRotaryEmbedding
 from sglang.srt.utils import get_compiler_backend, is_hpu
@@ -51,20 +52,6 @@ if TYPE_CHECKING:
     from sglang.srt.speculative.eagle_utils import EagleDraftInput, EagleVerifyInput
     from sglang.srt.speculative.spec_info import SpeculativeAlgorithm
 
-_PAD_SLOT_ID = 0
-_PAD_BLOCK_ID = 0
-
-PREFILL_BUCKET_MIN = 512
-PREFILL_BUCKET_STEP = 512
-PREFILL_BUCKET_MAX = 4096
-
-DECODE_BLOCK_BUCKET_MIN = 128
-DECODE_BLOCK_BUCKET_STEP = 128
-DECODE_BLOCK_BUCKET_MAX = 4096
-
-DECODE_BATCH_BUCKET_MIN = 1
-DECODE_BATCH_BUCKET_STEP = 32
-DECODE_BATCH_BUCKET_MAX = 192
 
 class ForwardMode(IntEnum):
     # Extend a sequence. The KV cache of the beginning part of the sequence is already computed (e.g., system prompt).
@@ -141,43 +128,6 @@ class CaptureHiddenMode(IntEnum):
 
     def is_last(self):
         return self == CaptureHiddenMode.LAST
-
-import itertools
-import math
-from vllm_hpu_extension.bucketing import find_bucket
-from vllm_hpu_extension.ops import batch2block, block2batch
-
-from vllm.utils import make_tensor_with_pad
-def flatten(in_list):
-    return list(itertools.chain(*in_list))
-
-def make_cpu_tensor(data, max_len, pad, dtype, flat):
-    if flat:
-        data = [flatten(data)]
-    result = make_tensor_with_pad(data,
-                                  max_len=max_len,
-                                  pad=pad,
-                                  dtype=dtype,
-                                  device='cpu')
-    return result
-
-def gather_list(tensor, indices, pad_value):
-    result = [pad_value] * len(indices)
-    for i, idx in enumerate(indices):
-        if idx is not None:
-            result[i] = tensor[idx]
-    return result
-
-
-def round_up(value: int, k: int) -> int:
-    return (value + k - 1) // k * k
-
-
-def pad_list(input, k, v):
-    input_len = len(input)
-    target_len = round_up(input_len, k)
-    padding = target_len - input_len
-    return input + [v] * padding
 
 
 @dataclass
@@ -294,81 +244,6 @@ class ForwardBatch:
     seq_idx: Optional[torch.Tensor] = None
     use_contiguous_pa: bool = True
 
-    @classmethod
-    def _set_block_mapping(cls, metadata, batch_size, device, dtype):
-        """Set block mapping using one-hot encoding of block groups."""
-
-        mask = torch.arange(0,
-                            metadata.page_size,
-                            device=device,
-                            dtype=torch.int32).unsqueeze(0)
-        mask = mask >= metadata.block_usage.unsqueeze(-1)
-        attn_bias = (torch.zeros_like(mask, dtype=dtype).masked_fill_(
-            mask, -math.inf))
-
-        # Handle out of bounds classes on CPU
-        block_groups = metadata.block_groups.to(torch.long)
-        block_mapping = torch.nn.functional.relu(block_groups)
-        block_mapping = torch.nn.functional.one_hot(block_mapping, num_classes=batch_size)
-        oob_values = block_groups.lt(0)
-        block_mapping.masked_fill_(oob_values.unsqueeze(-1), 0)
-        block_groups.masked_fill_(oob_values, batch_size)
-        return block_mapping.to(dtype), attn_bias, block_groups
-
-    @classmethod
-    def _set_block_scales(cls, metadata, device):
-        """Set block scales using batch2block and block2batch operations."""
-        block_mapping = metadata.block_mapping
-        ones = torch.ones((block_mapping.size(0),), device=device, dtype=block_mapping.dtype)
-        sums = batch2block(block2batch(ones, block_mapping), block_mapping)
-        block_scales = torch.reciprocal(torch.maximum(ones, sums))
-        return block_scales
-
-    @classmethod
-    def _init_block_metadata(cls, ret, model_runner, block_tables, slot_mapping, block_size):
-        """Initialize block metadata for HPU paged attention."""
-        device = "cpu"
-        dtype = model_runner.dtype
-
-        # Calculate block metadata
-        last_block_usage = [
-            slot % block_size + 1 for slot in slot_mapping
-        ]
-        block_groups = [[i] * len(bt) for i, bt in enumerate(block_tables)]
-        block_usage = [[block_size] * (len(bt) - 1) + [lbu]
-                      for bt, lbu in zip(block_tables, last_block_usage)
-                      if bt]
-        block_list = flatten(block_tables)
-        block_groups = flatten(block_groups)
-        block_usage = flatten(block_usage)
-        assert len(block_list) == len(block_groups)
-        assert len(block_list) == len(block_usage)
-
-        if ret.use_contiguous_pa:
-            # Pad block metadata if needed
-            block_bucket_size = max(max(block_list) + 1, len(block_list))
-            block_bucket_size = find_bucket(block_bucket_size, (DECODE_BLOCK_BUCKET_MIN, DECODE_BLOCK_BUCKET_STEP, DECODE_BLOCK_BUCKET_MAX))
-            indices = [None] * block_bucket_size
-            for i, bid in enumerate(block_list):
-                indices[bid] = i
-            padding_fn = lambda tensor, pad_value: gather_list(tensor, indices, pad_value)
-        else:
-            block_bucket_size = find_bucket(len(block_list), (DECODE_BLOCK_BUCKET_MIN, DECODE_BLOCK_BUCKET_STEP, DECODE_BLOCK_BUCKET_MAX))
-            padding_fn = lambda tensor, pad_value: pad_list(tensor, block_bucket_size, pad_value)
-
-        block_list = padding_fn(block_list, _PAD_BLOCK_ID)
-        block_groups = padding_fn(block_groups, -1)
-        block_usage = padding_fn(block_usage, 1)
-
-        # Convert to tensors
-        ret.block_list = torch.tensor(block_list, dtype=torch.long, device=device)
-        ret.block_groups = torch.tensor(block_groups, dtype=torch.long, device=device)
-        ret.block_usage = torch.tensor(block_usage, dtype=dtype, device=device)
-
-        # Set block mapping and scales
-        ret.block_mapping, ret.attn_bias, ret.block_groups = cls._set_block_mapping(ret, ret.batch_size, device, dtype)
-        ret.block_scales = cls._set_block_scales(ret, device)
-
 
     @classmethod
     def init_new(
@@ -376,7 +251,8 @@ class ForwardBatch:
         batch: ModelWorkerBatch,
         model_runner: ModelRunner,
     ):
-        device = model_runner.device
+        # device = model_runner.device
+        device = "cpu"
         extend_input_logprob_token_ids_gpu = None
         if batch.extend_input_logprob_token_ids is not None:
             extend_input_logprob_token_ids_gpu = (
@@ -442,30 +318,9 @@ class ForwardBatch:
 
         # Init position information
         if ret.forward_mode.is_decode():
-            if ret.positions is None:
-                ret.positions = clamp_position(batch.seq_lens)
             if ret.decode_seq_lens_cpu is None:
                 ret.decode_seq_lens_cpu = batch.decode_seq_lens
         else:
-            ret.extend_seq_lens = torch.tensor(
-                batch.extend_seq_lens, dtype=torch.int32
-            ).to(device, non_blocking=True)
-            ret.extend_prefix_lens = torch.tensor(
-                batch.extend_prefix_lens, dtype=torch.int32
-            ).to(device, non_blocking=True)
-            if model_runner.server_args.attention_backend not in ["torch_native", "hpu"]:
-                ret.extend_num_tokens = batch.extend_num_tokens
-                positions, ret.extend_start_loc = compute_position_triton(
-                    ret.extend_prefix_lens,
-                    ret.extend_seq_lens,
-                    ret.extend_num_tokens,
-                )
-            else:
-                positions, ret.extend_start_loc = compute_position_torch(
-                    ret.extend_prefix_lens, ret.extend_seq_lens
-                )
-            if ret.positions is None:
-                ret.positions = positions
             ret.extend_prefix_lens_cpu = batch.extend_prefix_lens
             ret.extend_seq_lens_cpu = batch.extend_seq_lens
             ret.extend_logprob_start_lens_cpu = batch.extend_logprob_start_lens
@@ -477,81 +332,25 @@ class ForwardBatch:
         if model_runner.server_args.lora_paths is not None:
             model_runner.lora_manager.prepare_lora_batch(ret)
 
-        seq_len_list = ret.extend_seq_lens_cpu
-        if model_runner.server_args.attention_backend == "hpu":
-            ret.page_size = model_runner.token_to_kv_pool_allocator.page_size
-            if ret.forward_mode.is_extend():
-                sum_seq_len = sum(seq_len_list)
-                max_prompt_len = find_bucket(sum_seq_len, (PREFILL_BUCKET_MIN, PREFILL_BUCKET_STEP, PREFILL_BUCKET_MAX))
-                ret.attn_bias, ret.seq_pos, ret.seq_idx = cls.make_hpu_attn_bias(
-                    seq_lens=seq_len_list,
-                    max_prompt_len=max_prompt_len,
-                    dtype=model_runner.dtype,
-                )
-                padding_len = max_prompt_len - sum_seq_len
-                max_prefill_seqs = model_runner.server_args.max_running_requests
-                ret.input_ids = torch.nn.functional.pad(ret.input_ids, (0, padding_len), value=0)
-                ret.positions = torch.nn.functional.pad(ret.positions, (0, padding_len), value=0)
-                ret.valid_seq_len = torch.tensor(sum_seq_len, dtype=torch.int32)
-                ret.extend_seq_lens = torch.nn.functional.pad(ret.extend_seq_lens, (0, max_prefill_seqs - ret.batch_size), value=0)
-                ret.out_cache_loc = torch.nn.functional.pad(ret.out_cache_loc, (0, padding_len), value=0)
-                ret.real_batch_size = ret.batch_size
-                ret.batch_size = 1
-            else:
-                ret.use_contiguous_pa = os.environ.get('SGLANG_HPU_CONTIGUOUS_PA',
-                                                'true').lower() in ['true', '1']
-                # Initialize block metadata for HPU paged attention
-                from sglang.srt.mem_cache.paged_allocator import HPUPagedTokenToKVPoolAllocator
-                paged_allocator: HPUPagedTokenToKVPoolAllocator = model_runner.token_to_kv_pool_allocator
-                padded_batch_size = find_bucket(ret.batch_size, (DECODE_BATCH_BUCKET_MIN, DECODE_BATCH_BUCKET_STEP, DECODE_BATCH_BUCKET_MAX))
-                block_tables = []
-                for i in range(ret.batch_size):
-                    block_tables.append(paged_allocator.block_manager.seq_info[ret.req_pool_indices[i].item()][0])
-
-                for i in range(padded_batch_size - ret.batch_size):
-                    block_tables.append([_PAD_BLOCK_ID])
-
-                padding_len = padded_batch_size - ret.batch_size
-                input_ids = torch.nn.functional.pad(ret.input_ids, (0, padding_len), value=0)
-                positions = torch.nn.functional.pad(ret.positions, (0, padding_len), value=0)
-                ret.valid_seq_len = torch.ones(padded_batch_size, dtype=torch.int32)
-                ret.out_cache_loc = torch.nn.functional.pad(ret.out_cache_loc, (0, padding_len), value=0)
-                ret.real_batch_size = ret.batch_size
-                
-                slot_mapping = ret.out_cache_loc
-                block_size = paged_allocator.page_size
-                ret.input_ids = input_ids
-                ret.positions = positions
-                ret.batch_size = padded_batch_size
-                cls._init_block_metadata(ret, model_runner, block_tables, slot_mapping, block_size)
-        return ret
-
-    @classmethod
-    def make_hpu_attn_bias(cls, seq_lens, max_prompt_len, dtype):
-        seq_pos = [list(range(sl)) for sl in seq_lens]
-        seq_idx = [[i] * sl for i, sl in enumerate(seq_lens)]
-        seq_pos = make_cpu_tensor(seq_pos,
-                                  max_len=max_prompt_len,
-                                  pad=-1,
-                                  dtype=torch.long,
-                                  flat=True)
-        seq_idx = make_cpu_tensor(seq_idx,
-                                  max_len=max_prompt_len,
-                                  pad=-1,
-                                  dtype=torch.long,
-                                  flat=True)
-        # q_seq_idx = seq_idx.unsqueeze(-1)
-        # kv_seq_idx = seq_idx.unsqueeze(-2)
-        # q_seq_pos = seq_pos.unsqueeze(-1)
-        # kv_seq_pos = seq_pos.unsqueeze(-2)
-        # seq_idx = q_seq_idx != kv_seq_idx
-        # seq_pos = kv_seq_pos > q_seq_pos
-        # attn_mask = seq_idx | seq_pos
         
-        # attn_bias.masked_fill_(attn_mask, -math.inf)
-        # return attn_bias.unsqueeze(1)
-        attn_bias = torch.zeros(1, 1, max_prompt_len, max_prompt_len, dtype=dtype)
-        return attn_bias, seq_pos, seq_idx
+        if model_runner.server_args.attention_backend == "hpu":
+            ret.positions = batch.positions
+            ret.batch_size = batch.batch_size
+            ret.extend_start_loc = batch.extend_start_loc
+            ret.page_size = batch.page_size
+            ret.attn_bias = batch.attn_bias
+            ret.seq_pos = batch.seq_pos
+            ret.seq_idx = batch.seq_idx
+            ret.valid_seq_len = batch.valid_seq_len
+            ret.extend_seq_lens = batch.extend_seq_lens_padded
+            ret.block_list = batch.block_list
+            ret.block_mapping = batch.block_mapping
+            ret.block_groups = batch.block_groups
+            ret.block_usage = batch.block_usage
+            ret.block_scales = batch.block_scales
+            ret.use_contiguous_pa = batch.use_contiguous_pa
+            ret.real_batch_size = batch.real_batch_size
+        return ret
 
     def get_merged_image_inputs(self) -> Optional[ImageInputs]:
         """
